refactor http communicator to use vql api and introduce certificate pinning
